requirements.txt
flask == 2.2.2
pandas == 1.4.2
openai==0.27.2
python-dotenv==1.0.0